src/main/java/cc/oakk/racing/domain/Position.java
package cc.oakk.racing.domain;

import java.util.Objects;

public class Position implements Comparable<Position> {
    private int position;

    public Position(int position) {
        if (position < 0) {
            throw new IllegalArgumentException("Given Position should be equals to 0 or greater than 0.");
        }
        this.position = position;
    }

    public void increase() {
        position++;
    }

    public int getPosition() {
        return position;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Position position1 = (Position) o;
        return position == position1.position;
    }

    @Override
    public int hashCode() {
        return Objects.hash(position);
    }

    @Override
    public int compareTo(Position opponent) {
        return this.getPosition() - opponent.getPosition();
    }
}
